Fixed routing overall. Now the root will redirect to home and the calc parent has it's own routeroutlet to show each calculator depending on menu button press.
