+++ src/app/app.component.html
@@ -13,12 +13,12 @@
         </a>
       </div>
       <div class="navButton">
-        <a [routerLink]="['./calc']">
+        <a [routerLink]="['./calc/plasticshrinkage']">
           <button mat-raised-button color="primary">Plastic Shrinkage Cracks</button>
         </a>
       </div>
       <div class="navButton">
-        <a [routerLink]="['./calc']">
+        <a [routerLink]="['./calc/tcpowers']">
           <button mat-raised-button color="primary">TC Powers Volume</button>
         </a>
       </div>
@@ -33,7 +33,7 @@
         </a>
       </div>
       <div class="navButton">
-        <a [routerLink]="['./calc']">
+        <a>
           <button mat-raised-button color="primary">Login</button>
         </a>
       </div>
