+++ src/app/main-content/calculator-parent/calculator-parent.component.html
@@ -1,35 +1,7 @@
 <p>calculator-parent works!</p>
-<div NgIf="CalcName == 'plasticshrinkage'">
     <a class="about-tutorial-and-calc-button">
         <button mat-raised-button color="basic">Calculator</button>
         <button mat-raised-button color="basic">About</button>
         <button mat-raised-button color="basic">Tutorial</button>
     </a>
-    <app-plastic-shrinkage-cracks></app-plastic-shrinkage-cracks>
-    <!-- insert select here -->
-</div>
-<div NgIf="CalcName == 'tcpowers'">
-    <a class="about-tutorial-and-calc-button">
-        <button mat-raised-button color="basic">Calculator</button>
-        <button mat-raised-button color="basic">About</button>
-        <button mat-raised-button color="basic">Tutorial</button>
-    </a>
-    <app-tcpowers></app-tcpowers>
-    <!-- insert select here -->
-</div>
-<div NgIf="CalcName == 'bogue'">
-    <a class="about-tutorial-and-calc-button">
-        <button mat-raised-button color="basic">Calculator</button>
-        <button mat-raised-button color="basic">About</button>
-        <button mat-raised-button color="basic">Tutorial</button>
-    </a>
-    <app-bogue></app-bogue>
-</div>
-<div NgIf="CalcName == 'morterandmix'">
-    <a class="about-tutorial-and-calc-button">
-        <button mat-raised-button color="basic">Calculator</button>
-        <button mat-raised-button color="basic">About</button>
-        <button mat-raised-button color="basic">Tutorial</button>
-    </a>
-    <app-mortar-and-mix></app-mortar-and-mix>
-</div>
+<router-outlet></router-outlet>
