+++ src/app/main-content/calculator-parent/calculator-parent.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { MortarAndMixComponent } from './mortar-and-mix/mortar-and-mix.component';
 import { BogueComponent } from './bogue/bogue.component';
 import { TcpowersComponent } from './tcpowers/tcpowers.component';
-import { Router, ActivatedRoute, ParamMap } from '@angular/router';
+import { Router, ActivatedRoute, RouterOutlet } from '@angular/router';
 import {MatDividerModule} from '@angular/material/divider';
 import {MatButtonModule} from '@angular/material/button';
 import { PlasticShrinkageCracksComponent } from './plastic-shrinkage-cracks/plastic-shrinkage-cracks.component';
@@ -11,7 +11,15 @@ import { PlasticShrinkageCracksComponent } from './plastic-shrinkage-cracks/plas
 @Component({
   selector: 'app-calculator-parent',
   standalone: true,
-  imports: [MortarAndMixComponent, BogueComponent, TcpowersComponent, PlasticShrinkageCracksComponent, MatDividerModule, MatButtonModule],
+  imports: [
+    MortarAndMixComponent, 
+    BogueComponent, 
+    TcpowersComponent, 
+    PlasticShrinkageCracksComponent, 
+    MatDividerModule, 
+    MatButtonModule,
+    RouterOutlet
+  ],
   templateUrl: './calculator-parent.component.html',
   styleUrl: './calculator-parent.component.css'
 }
