--- src/app/main-content/main-content.component.html
@@ -1,4 +1,0 @@
-<p></p>
-<div>
-    <router-outlet></router-outlet>
-</div>
